statement: move exception to calculator
