+++ calculator/src/main/java/ru/neoflex/calculator/dto/exception/PrescoringFailedException.java
@@ -1,4 +1,4 @@
-package ru.neoflex.deal.exception;
+package ru.neoflex.calculator.dto.exception;
 
 public class PrescoringFailedException extends RuntimeException {
