+++ deal/src/main/java/ru/neoflex/deal/service/CalculatorApiService.java
@@ -10,7 +10,7 @@ import ru.neoflex.calculator.dto.offer.request.LoanStatementRequestDto;
 import ru.neoflex.calculator.dto.offer.response.LoanOfferDto;
 import ru.neoflex.calculator.dto.scoring.request.ScoringDataDto;
 import ru.neoflex.calculator.dto.scoring.response.CreditDto;
-import ru.neoflex.deal.exception.PrescoringFailedException;
+import ru.neoflex.calculator.dto.exception.PrescoringFailedException;
 
 import java.util.Arrays;
 import java.util.List;
